+++ changelog.md
@@ -7,3 +7,4 @@
 - Added the ability to create a new workout file
 - Added the ability to select a stored workout and view it
 - Added error message when starting an empty workout
+- redundant `view workout` button removed